implemented bone item, restructured code
the bone does not yet work as intended, spawning only 1 bone at a time. _draw and _update are now at the end of my code since they can only be called once. all the tables and function declarations are now at the beginning of my code and everything that needs to happen every frame (checking how many bones there are or whether my character needs to be teleported) is now within the update function. the game still works, which is a miracle!